requirements.txt
flask
flake8<4.0
pytesseract>=0.2.1,<0.3
pillow # not directly required, pinned by Snyk to avoid a vulnerability
opencv-python-headless
#opencv-python < 5.0
numpy < 2.0
azure-storage-blob
pyyaml